add delete features, loading and sweet alert

+++ src/script/view/home.js
@@ -1,10 +1,13 @@
 import Utils from "../utils.js";
 import NotesApi from "../data/remote/notes-api.js";
+import Swal from "sweetalert2";
 
+const noteListContainer = document.querySelector("#notesContainer");
 const listElement = document.querySelector("note-list");
 const notesTitleElement = document.querySelector("#notes-title");
 const notesBodyElement = document.querySelector("#notes-body");
 const formNotesElement = document.querySelector("#form-notes");
+const loadingElement = document.querySelector("indikator-loading");
 
 const home = () => {
   getNotes();
@@ -13,9 +16,13 @@ const home = () => {
 };
 
 const getNotes = () => {
+  showLoading();
   NotesApi.getNotes()
-    .then((result) => displayResult(result))
-    .catch((error) => console.log(error));
+    .then((result) => {
+      displayResult(result);
+      showNoteList();
+    })
+    .catch((error) => showError(error));
 };
 
 const displayResult = (notes) => {
@@ -59,7 +66,7 @@ const saveNotes = () => {
         showMessage(result);
         getNotes();
       })
-      .catch((error) => console.log(error));
+      .catch((error) => showError(error));
 
     notesTitleElement.value = "";
     notesBodyElement.value = "";
@@ -68,17 +75,53 @@ const saveNotes = () => {
 
 const deleteNotes = () => {
   document.addEventListener("delete-note", (event) => {
-    NotesApi.deleteNotes(event.detail.id)
-      .then((result) => {
-        showMessage(result);
-        getNotes();
-      })
-      .catch((error) => console.log(error));
+    Swal.fire({
+      icon: "question",
+      title: "Are you sure",
+      text: "You want to delete this note",
+      showCancelButton: true,
+      footer: "<p>Dicoding notes &copy; 2025</p>",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        NotesApi.deleteNotes(event.detail.id)
+          .then((result) => {
+            showMessage(result);
+            getNotes();
+          })
+          .catch((error) => showError(error));
+      }
+    });
   });
 };
 
 const showMessage = (message = "Check your internet connection") => {
-  alert(message);
+  Swal.fire({
+    title: message,
+    icon: "success",
+  });
+};
+
+const showError = (error) => {
+  Swal.fire({
+    icon: "error",
+    title: "Oops...",
+    text: error.message,
+    footer: "Dicoding notes &copy; 2025",
+  });
+};
+
+const showLoading = () => {
+  Array.from(noteListContainer.children).forEach((element) => {
+    Utils.hideElement(element);
+  });
+  Utils.showElement(loadingElement);
+};
+
+const showNoteList = () => {
+  Array.from(noteListContainer.children).forEach((element) => {
+    Utils.hideElement(element);
+  });
+  Utils.showElement(listElement);
 };
 
 export default home;
